chore: 2025-05-04

+++ automl/src/main/java/org/xiaoshuyui/automl/module/tool/entity/PredictRequest.java
@@ -7,6 +7,6 @@ import lombok.Data;
 public class PredictRequest {
   private String image_data;
   private List<String> classes;
-  private int model_id;
+  private long model_id;
   private String prompt;
 }
